CVE-2016-1238: avoid loading optional modules from .
The change to Encode.pm is the most critical part of this patch.

Without this change, and process that uses Encode started with a
current directory that's world writable (such as /tmp) and if there's
no global Encode::ConfigLocal, can be attacked by another user
by creating /tmp/Encode/ConfigLocal.pm

It's possible most of the tools changed here do not need to be updated,
but I chose a conservative path.