Encode::MIME::Header: decode: Rewrite decoder and implement full support for check flags
Now decoder fully respect check flags and so it does not croak anymore when
it is called without any flag (or Encode::FB_DEFAULT flag) and when in
input string is invalid or incorrect charset. It means that implementation
of MIME decoder should match behaviour described in Encode module
documentation.

Invalid and unknown MIME words are not touched and they stay in decoded
output as were on input.

Previous behaviour (croak on unknown charset) can be achieved by passing
Encode::FB_CROAK check flag to decode().