Fix more XS problems in Encode.xs file
With CODE: directive use void return value only for subroutines which do
not return a value. And do not set ST(0) directly. This is described in
perlxs documentation.

Use PREINIT: directive for variable declarations and INIT: for initial
assignment. Due to C89 compatibility variable declarations needs to be done
at the beginning of the block

Remove blocks in directives. After above fixes they are not needed anymore.