Fix https://rt.cpan.org/Ticket/Display.html?id=124399
Encode 2.95 broke this module, which is testing for a particular warning
message.  It was not my intent to change any part of the API with the
commits that comprise 2.95, and it turns out that the message is
different only in one or two perl versions where an interaction with a
deprecation caused the new message to be displayed.

The commit here just keeps the old code being used into the 5.27 series.
It probably could work starting in 5.27.2, but its easier to use a
marker of the existence of a newer function, and since the affected
versions are all development releases, I went with the marker.