Include 5.28 core UTF8 to/from U32/U64 functions
This patch copies the perl core's functions that translate between UTF-8
and the equivalent code point.  It strips off functionality that isn't
likely to ever be needed by Encode.  These functions were enhanced in
the 5.27 development cycle to meet special needs of Encode.

Doing this allows Encode to have the same behavior on any perl that it
can be used with.  It also fixes various potential security issues when
Encode is used on early perls.  This is because the earlier translation
function to/from UTF-8, weren't really aware of the potential security
issues that have been fixed over the years.  By using the functions now
bundled with Encode, these defective early functions are bypassed.